Unique QID, Deleteable new questions
QID's should be unique now, if inefficient.
When we do not have a prompt displayed that we can actually delete, the delete button will be greyed out, but the moment that we can, it will be clickable.
We can now delete and click on questions in prompt history even if they are created on the fly.
The code to add a listener to a recentQuestion button is now a separate method we can use in different places