the server shouldl be worked for handling
the server should be worked for handling login and create request.

when a login button is clicked and both fields of account and password is not empty, then it will sends a login request to the server, along with informatioin about email, password, autoLogin in requestbody. Then the handler using these information to find if they are in MangoDB, and send back the login status to loginScreen. If loged in successfully, then open SayIt and close loginScreen.

Co-Authored-By: Hannes <hanisphy@gmail.com>